refactor(web): extract shared timeAgo helper (Epic B)
timeAgo was duplicated in DiffView, ReviewDetail, and ReviewList — and had
drifted (ReviewList returned bare values + appended ' ago' at the call site,
the others embedded ' ago'). Extract one lib/time.ts and use it everywhere.
