+++ web/src/components/review/DiffView.tsx
@@ -2,6 +2,7 @@ import { useState, useCallback } from 'react';
 import { MessageSquare, Check, ChevronDown, ChevronRight, Send, X, FileText } from 'lucide-react';
 import type { FileDiff, DiffHunk, DiffLine, ReviewComment } from '../../api';
 import { C, fonts } from '@/lib/design';
+import { timeAgo } from '../../lib/time';
 
 const mono = fonts.mono;
 
@@ -370,18 +371,6 @@ function FileDiffCard({
       )}
     </div>
   );
-}
-
-function timeAgo(iso: string): string {
-  const then = new Date(iso).getTime();
-  if (Number.isNaN(then)) return '';
-  const secs = Math.max(0, Math.floor((Date.now() - then) / 1000));
-  if (secs < 60) return `${secs}s ago`;
-  const mins = Math.floor(secs / 60);
-  if (mins < 60) return `${mins}m ago`;
-  const hrs = Math.floor(mins / 60);
-  if (hrs < 24) return `${hrs}h ago`;
-  return `${Math.floor(hrs / 24)}d ago`;
 }
 
 /* ── Exports ── */
