+++ web/src/components/review/ReviewDetail.tsx
@@ -6,18 +6,7 @@ import {
 } from '../../api';
 import { DiffView } from './DiffView';
 import { C, fonts } from '@/lib/design';
-
-function timeAgo(iso: string): string {
-  const then = new Date(iso).getTime();
-  if (Number.isNaN(then)) return '';
-  const secs = Math.max(0, Math.floor((Date.now() - then) / 1000));
-  if (secs < 60) return `${secs}s ago`;
-  const mins = Math.floor(secs / 60);
-  if (mins < 60) return `${mins}m ago`;
-  const hrs = Math.floor(mins / 60);
-  if (hrs < 24) return `${hrs}h ago`;
-  return `${Math.floor(hrs / 24)}d ago`;
-}
+import { timeAgo } from '../../lib/time';
 
 const statusBadge: Record<string, { bg: string; fg: string; label: string }> = {
   pending: { bg: C.amberSoft, fg: C.amber, label: 'Review needed' },
